# beckn-np-onboarding/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description:  ONDC Network Participant Onboarding APIs
  version: 0.0.6
  title: APIs to be used by ONDC Network Participant 
  # put the contact info for your development or API team
  contact:
    email: team@ondc.org

  license:
    name: TBD
    url: https://ondc.org

# tags are used for organizing operations
...
paths:
...
  /signature/generate:
    post:
      tags:
      - Creating test encryption keys and signing keys and also encryption and signing API
      summary: "Generate signature"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      operationId: signature
      parameters:
      - in: body
        name: Generate signature Request 
        description: Generate signature for requested data
        schema:
          $ref: '#/definitions/RequestSigData'
        required: true
      responses:
        "200":
          description: OK
          schema:
            type: string
            example : signature
  /signature/verify:
    post:
      tags:
      - Creating test encryption keys and signing keys and also encryption and signing API
      summary: "Signature Verify"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      operationId: signatureVerify
      parameters:
      - in: body
        name: Signature Key Verification Request 
        description: Verifies whether the given Signature is Valid
        schema:
          $ref: '#/definitions/RequestData'
      responses:
        "200":
          description: OK
          schema:
            type: string
            example : Signature is valid
        "500":
          description: OK
          schema:
            type: string
            example : Signature is not valid
  /signature/generate/onboarding:
    post:
      tags:
      - Creating test encryption keys and signing keys and also encryption and signing API
      summary: "Generate Signature"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      operationId: signaturePlan
      parameters:
      - in: body
        name: Generate signature Request 
        description: Generate signature for requested data
        schema:
          $ref: '#/definitions/RequestSigData'
        required: true
      responses:
        "200":
          description: OK
          schema:
            type: string
            example: signature
  /signature/verify/onboarding:
    post:
      tags:
      - Creating test encryption keys and signing keys and also encryption and signing API
      summary: "Verify Signature"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      operationId: signatureVerifyPlan
      parameters:
      - in: body
        name: Signature Key Verification Request 
        description: Verifies whether the given Signature is Valid
        schema:
          $ref: '#/definitions/RequestData'
        required: true
      responses:
        "200":
          description: OK
          schema:
            type: string
            example : Signature is valid
        "500":
          description: OK
          schema:
            type: string
            example : Signature is not valid
  /ocsp/verify:
    post:
      tags:
      - Creating test encryption keys and signing keys and also encryption and signing API
      summary: "Verify SSL URL using OCSP"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      operationId: ocspVerify
      parameters:
      - in: body
        name: SSL URL Request 
        description: Verifies whether the given URL SSL is Valid
        schema:
          $ref: '#/definitions/Web'
        required: true
      responses:
        "200":
          description: OK
          schema:
            type: string
            example: Valid | Invalid
  /encrypt/text:
    post:
      tags:
      - Creating test encryption keys and signing keys and also encryption and signing API
      summary: "Encrypt Text"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      operationId: encryptText
      parameters:
      - in: body
        name: Encryption Request 
        description: Encrypt the given Text
        schema:
          $ref: '#/definitions/RequestEncryptData'
        required: true
      responses:
        "200":
          description: OK
          schema:
            type: string
            example: Encrypt Text
  /decrypt/text:
    post:
      tags:
      - Creating test encryption keys and signing keys and also encryption and signing API
      summary: "Decrypt encrypted Text"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      operationId: decryptText
      parameters:
      - in: body
        name: Decryption Request 
        description: Verifies whether the given Signature is Valid
        schema:
          $ref: '#/definitions/RequestDecryptData'
        required: true
      responses:
        "200":
          description: OK
          schema:
            type: string
            example: Decrypt Text
...
definitions:
...
  RequestData:
      type: object
      properties:
        sign:
          type: string
          format: string
          description: Signature
          example: 8B5DLTKkm/0dP+xXZ/1RlbSEgbg+4fgU1kn+aVFJasR2OnFLmizJmFSswcDl1IBRmyrNPfTZZVbl++biljrqAg==
        requestData:
          type: string
          format: string
          description: Value
          example: Text data for signature
        clientPublicKey:
          type: string
          format: string
          description: Public key
          example: MCowBQYDK2VuAyEAIdWagNWnmgro84jyJv/y21WRqgSeL05ZcUsuAMVJC3Y=
  RequestSigData:
      type: object
      properties:
        requestData:
          type: string
          format: string
          description: Value
          example: Text data for signature
        clientPrivateKey:
          type: string
          format: string
          description: Private key
          example: MFECAQEwBQYDK2VuBCIEIIiPHne7kFP5L8Ju4M6vje57LZOVoUo6qatNxQlzE4BcgSEAIdWagNWnmgro84jyJv/y21WRqgSeL05ZcUsuAMVJC3Y=
...
  Web:
      type: object
      properties:
        domain:
          type: string
          format: string
          description: Web domain for OCSP verification
          example: https://nsdl.co.in
  RequestEncryptData:
      type: object
      properties:
        clientPrivateKey:
          type: string
          format: string
          description: Network Particiant private key
          example: MFECAQEwBQYDK2VuBCIEIIiPHne7kFP5L8Ju4M6vje57LZOVoUo6qatNxQlzE4BcgSEAIdWagNWnmgro84jyJv/y21WRqgSeL05ZcUsuAMVJC3Y=
        proteanPublicKey:
          type: string
          format: string
          description: Server public key
          example: MFECAQEwBQYDK2VuBCIEINDj+FYrFzcQkKlvqPzARtaYzV+A229aCOII51IkhCxkgSEAlfIoajvDFLfZ/4ysTEscSsKJoW9NkGCi+f4QP3vzhBA=
        value:
          type: string
          format: string
          description: Network Particiant public key
          example: Text
  RequestDecryptData:
      type: object
      properties:
        clientPublicKey:
          type: string
          format: string
          description: Network Particiant public key
          example: MCowBQYDK2VuAyEAIdWagNWnmgro84jyJv/y21WRqgSeL05ZcUsuAMVJC3Y=
        proteanPrivateKey:
          type: string
          format: string
          description: Server private key
          example: MCowBQYDK2VuAyEAlfIoajvDFLfZ/4ysTEscSsKJoW9NkGCi+f4QP3vzhBA=
        value:
          type: string
          format: string
          description: Network Particiant public key
          example: Encrypt Text        
    
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /sujitrs/OndcNP/0.0.3
schemes:
 - https